Refactor questions domain code into Survey objects
